lambda_ingest/requirements.txt
psycopg2-binary==2.9.7
boto3==1.28.62
openai>=1.0.0